Move the updater into its own folder
packages/app/updater.ts and the channel helpers it depends on followed no
shape: the Electron-bound class sat at the top level while its one pure
module and that module's test lived in lib/, named update-channel after
the folder they had been exiled to rather than after what they are.

They now sit together on the packages/app/blocker pattern — index.ts for
the code bound to Electron, pure modules beside it under plain names with
their sibling tests:

  updater.ts                  -> updater/index.ts
  lib/update-channel.ts       -> updater/channel.ts
  lib/update-channel.test.ts  -> updater/channel.test.ts

A pure rename, done with git mv so history follows. Nothing outside the
folder imported either lib module, and @/updater and ./updater resolve to
index.ts unchanged, so no call site moves with them; the only edits are
the relative imports that crossed the new folder boundary and one comment
naming the file beside it.

No changelog line: nothing a user of the app can see changes.
