+++ packages/app/updater/index.ts
@@ -1,10 +1,10 @@
 import { is } from "@electron-toolkit/utils";
 import { autoUpdater } from "electron-updater";
 import { config } from "@/config";
-import { ipc } from "./ipc";
-import { log } from "./lib/log";
-import { resolveUpdateChannel } from "./lib/update-channel";
-import { main } from "./main";
+import { ipc } from "../ipc";
+import { log } from "../lib/log";
+import { main } from "../main";
+import { resolveUpdateChannel } from "./channel";
 
 class AppUpdater {
   private applyChannel() {
